+++ R/ReadNetCDF.R
@@ -4,7 +4,11 @@
 #' over using \code{\link[ncdf4]{ncvar_get}} is that the output is a tidy data.table
 #' with proper dimensions.
 #'
-#' @param file file to read from.
+#' @param file source to read from. Must be one of:
+#'    * A string representing a local file with read access.
+#'    * A string representing a URL readable by [ncdf4::nc_open()].
+#'      (this includes DAP urls).
+#'    * A netcdf object returned by [ncdf4::nc_open()].
 #' @param vars a character vector with the name of the variables to read. If
 #' \code{NULL}, then it read all the variables.
 #' @param out character indicating the type of output desired
@@ -89,17 +93,36 @@
 #' file with a nice printing method.
 #'
 #' @examples
-#' \dontrun{
-#' file <- "file.nc"
+#' file <- system.file("extdata", "temperature.nc", package = "metR")
 #' # Get a list of variables.
 #' variables <- GlanceNetCDF(file)
 #' print(variables)
+#'
+#' # The object returned by GlanceNetCDF is a list with lots
+#' # of information
+#' str(variables)
 #'
 #' # Read only the first one, with name "var".
 #' field <- ReadNetCDF(file, vars = c(var = names(variables$vars[1])))
 #' # Add a new variable.
 #' # ¡Make sure it's on the same exact grid!
-#' field[, var2 := ReadNerCDF(file2, out = "vector", subset = list(lat = 90:10))]
+#' field[, var2 := ReadNetCDF(file, out = "vector")]
+#'
+#' \dontrun{
+#' # Using a DAP url
+#' url <- "http://iridl.ldeo.columbia.edu/SOURCES/.Models/.SubX/.GMAO/.GEOS_V2p1/.hindcast/.ua/dods"
+#' field <- ReadNetCDF(url, subset = list(M = 1,
+#'                                        P = 10,
+#'                                        S = "1999-01-01"))
+#'
+#' # In this case, opening the netcdf file takes a non-neglible
+#' # amount of time. So if you want to iterate over many dimensions,
+#' # then it's more efficient to open the file first and then read it.
+#'
+#' ncfile <- ncdf4::nc_open(url)
+#' field <- ReadNetCDF(ncfile, subset = list(M = 1,
+#'                                        P = 10,
+#'                                        S = "1999-01-01"))
 #' }
 #'
 #' @export
@@ -116,8 +139,12 @@ ReadNetCDF <- function(file, vars = NULL,
 
     out <- out[1]
     checks <- makeAssertCollection()
-    assertCharacter(file, len = 1, min.chars = 1, any.missing = FALSE, add = checks)
-    assertAccess(file, "r", add = checks)
+
+    if (!inherits(file, "ncdf4")) {
+        assertCharacter(file, len = 1, min.chars = 1, any.missing = FALSE, add = checks)
+        assertURLFile(file, add = checks)
+    }
+
     assertCharacter(vars, null.ok = TRUE, any.missing = FALSE, unique = TRUE,
                     add = checks)
     assertChoice(out, c("data.frame", "vector", "array", "vars"), add = checks)
@@ -127,23 +154,22 @@ ReadNetCDF <- function(file, vars = NULL,
 
     reportAssertions(checks)
 
-    ncfile <- ncdf4::nc_open(file)
+    if (!inherits(file, "ncdf4")) {
+        ncfile <- ncdf4::nc_open(file)
+        on.exit({
+            ncdf4::nc_close(ncfile)
+        })
+    } else {
+       ncfile <- file
+    }
+
 
     if (out[1] == "vars") {
         r <- list(vars = ncfile$var,
                   dims = ncfile$dim)
         class(r) <- c("nc_glance", class(r))
-        # options(OutDec = dec)
         return(r)
     }
-
-    dec <- getOption("OutDec")
-    # Dejemos todo prolijo antes de salir.
-    options(OutDec = ".")
-    on.exit({
-        options(OutDec = dec)
-        ncdf4::nc_close(ncfile)
-    })
 
     if (is.null(vars)) {
         vars <- names(ncfile$var)
@@ -174,11 +200,6 @@ ReadNetCDF <- function(file, vars = NULL,
     }
     names(dims) <- ids
 
-    # if (out[1] == "vars") {
-    #     r <- list(vars = unname(vars), dimensions = dimensions, dims = dims)
-    #     # options(OutDec = dec)
-    #     return(r)
-    # }
     ## Hago los subsets
     # Me fijo si faltan dimensiones
     subset <- .expand_chunks(subset)